requirements.txt
pandas~=2.0.3
yaml~=0.2.5
pyyaml~=6.0
numpy~=1.25.2
torchvision~=0.15.2
timm~=0.9.2
easydict~=1.10
pillow~=9.4.0
torchtoolbox~=0.1.8.2